refactor: combine process column code
